+++ documentation/playground/src/base/IconExamples.tsx
@@ -9,6 +9,7 @@ export function IconExamples() {
         <Icon name="alert-diamond" />
         <Icon name="mouse" screenReaderText="Click to connect your mouse" />
         <Icon thin name="ai-mouse" screenReaderText="Click to connect your mouse" />
+        <Icon name="nothingtoseehere" />
       </div>
     </div>
   );